cart

+++ src/main/java/dev/mananhemani/markethub/Controllers/CartController.java
@@ -17,7 +17,7 @@ public class CartController {
     @PostMapping("/carts/products/{productId}/quantity/{quantity}")
     public ResponseEntity<CartDTO> addProductToCart(
             @PathVariable Long productId,
-            @PathVariable Long quantity){
+            @PathVariable Integer quantity){
 
         CartDTO cartDTO = cartService.addProductToCart(productId,quantity);
