+++ src/main/java/dev/mananhemani/markethub/Services/Cart/CartService.java
@@ -4,5 +4,5 @@ import dev.mananhemani.markethub.DTOs.Cart.CartDTO;
 
 
 public interface CartService {
-    CartDTO addProductToCart(Long productId, Long quantity);
+    CartDTO addProductToCart(Long productId, Integer quantity);
 }
